исправлены замечания после ревью
