+++ src/test/java/com/example/AnimalTestGetFoodWithParameters.java
@@ -1,0 +1,47 @@
+package com.example;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+@RunWith(Parameterized.class)
+public class AnimalTestGetFoodWithParameters {
+    private Animal animal;
+    private static final String ANIMAL_KIND_HERBIVORE = "Травоядное";
+    private static final String ANIMAL_KIND_PREDATOR = "Хищник";
+    private static final List<String> FOODS_HERBIVORE = List.of("Трава", "Различные растения");
+    private static final List<String> FOODS_PREDATOR = List.of("Животные", "Птицы", "Рыба");
+
+    private String animalKind;
+    private List<String> foods;
+
+    public AnimalTestGetFoodWithParameters(String animalKind, List<String> foods) {
+        this.animalKind = animalKind;
+        this.foods = foods;
+    }
+
+    @Parameterized.Parameters
+    public static Object[][] getParameters() {
+        return new Object[][]{
+                {ANIMAL_KIND_HERBIVORE, FOODS_HERBIVORE},
+                {ANIMAL_KIND_PREDATOR, FOODS_PREDATOR}
+        };
+    }
+
+    @Before
+    public void setUp() {
+        animal = new Animal();
+    }
+
+    @Test
+    public void testGetFood() throws Exception {
+        List<String> actual = animal.getFood(animalKind);
+        assertEquals("Список еды для животных не совпадает с ожидаемым",
+                    foods, actual);
+    }
+}
